Move CredentialInputAdapter into MainActivity and make it private
This also fixes a thinko with ftps URLs where the protocal length
wasn't taken into account properly.
